Allow for additional Javascript engines (#344)
Moved all Nodejs work into `js.nodejs` subpackages.
Created some base classes that are engine-agnostic in
the `js.base` subpackage.
