+++ asciidoctor-gradle-js/src/main/groovy/org/asciidoctor/gradle/js/nodejs/AsciidoctorNodeJSBasePlugin.groovy
@@ -13,26 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.asciidoctor.gradle.js
+package org.asciidoctor.gradle.js.nodejs
 
 import groovy.transform.CompileStatic
+import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.Task
-import org.ysb33r.gradle.nodejs.NodeJSExtension
 
-/** An extension to configure Node.js.
- *
+/**
  * @since 3.0
  */
 @CompileStatic
-class AsciidoctorJSNodeExtension extends NodeJSExtension {
-    public final static String NAME = 'asciidoctorNodejs'
+class AsciidoctorNodeJSBasePlugin implements Plugin<Project> {
 
-    AsciidoctorJSNodeExtension(Project project) {
-        super(project)
-    }
+    public static final String TASK_GROUP = 'Documentation'
 
-    AsciidoctorJSNodeExtension(Task task) {
-        super(task, NAME)
+    @Override
+    void apply(Project project) {
+        project.extensions.create( AsciidoctorJSExtension.NAME, AsciidoctorJSExtension, project )
+        project.extensions.create( AsciidoctorJSNodeExtension.NAME, AsciidoctorJSNodeExtension, project )
+        project.extensions.create( AsciidoctorJSNpmExtension.NAME, AsciidoctorJSNpmExtension, project )
     }
 }
