+++ asciidoctor-gradle-js/src/main/groovy/org/asciidoctor/gradle/js/nodejs/internal/PackageDescriptor.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.asciidoctor.gradle.js.internal
+package org.asciidoctor.gradle.js.nodejs.internal
 
 import groovy.transform.CompileStatic
