+++ asciidoctor-gradle-js/src/main/resources/META-INF/gradle-plugins/org.asciidoctor.js.base.properties
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-implementation-class=org.asciidoctor.gradle.js.AsciidoctorJSBasePlugin
+implementation-class=org.asciidoctor.gradle.js.nodejs.AsciidoctorNodeJSBasePlugin
