+++ asciidoctor-gradle-js/src/main/resources/META-INF/gradle-plugins/org.asciidoctor.js.convert.properties
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-implementation-class=org.asciidoctor.gradle.js.AsciidoctorJSPlugin
+implementation-class=org.asciidoctor.gradle.js.nodejs.AsciidoctorNodeJSPlugin
