+++ asciidoctor-gradle-js/src/test/groovy/org/asciidoctor/gradle/js/nodejs/AsciidoctorTaskSpec.groovy
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.asciidoctor.gradle.js
+package org.asciidoctor.gradle.js.nodejs
 
-
+import org.asciidoctor.gradle.js.nodejs.AsciidoctorTask
 import org.gradle.api.Project
 import org.gradle.testfixtures.ProjectBuilder
 import spock.lang.Specification
